claudecode/requirements.txt
pygame==2.5.2
pyserial==3.5
PyQt5==5.15.10
numpy==1.24.3